test: separate old and new tracker e2e tests
make sure this is clear via the base classes and deprecated annotations
