+++ dhis-2/dhis-e2e-test/src/test/java/org/hisp/dhis/tracker/DeprecatedTrackerApiTest.java
@@ -25,71 +25,47 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.hisp.dhis.actions.tracker;
+package org.hisp.dhis.tracker;
 
-import java.util.List;
-
-import org.hisp.dhis.actions.MaintenanceActions;
+import org.hisp.dhis.ApiTest;
+import org.hisp.dhis.actions.LoginActions;
 import org.hisp.dhis.actions.RestApiActions;
-import org.hisp.dhis.dto.ApiResponse;
-
-import com.google.gson.JsonObject;
+import org.hisp.dhis.actions.metadata.ProgramActions;
+import org.hisp.dhis.actions.tracker.EventActions;
+import org.hisp.dhis.actions.tracker.TrackedEntityInstancesAction;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Tag;
 
 /**
  * @author Gintare Vilkelyte <vilkelyte.gintare@gmail.com>
+ *
  * @deprecated this is a base test class for "old" (deprecated) tracker which
  *             will be removed with "old" tracker
  */
 @Deprecated( since = "2.41" )
-public class EventActions
-    extends RestApiActions
+@Tag( "tracker" )
+@Tag( "category:tracker" )
+public class DeprecatedTrackerApiTest
+    extends ApiTest
 {
-    public EventActions()
+    protected RestApiActions enrollmentActions;
+
+    protected EventActions eventActions;
+
+    protected TrackedEntityInstancesAction trackedEntityInstancesAction;
+
+    protected LoginActions loginActions;
+
+    protected ProgramActions programActions;
+
+    @BeforeAll
+    public void beforeTracker()
     {
-        super( "/events" );
+        trackedEntityInstancesAction = new TrackedEntityInstancesAction();
+        loginActions = new LoginActions();
+        programActions = new ProgramActions();
+        eventActions = new EventActions();
+        enrollmentActions = new RestApiActions( "/enrollments" );
     }
 
-    /**
-     * Hard deletes event.
-     *
-     * @param eventId
-     * @return
-     */
-    @Override
-    public ApiResponse delete( String eventId )
-    {
-        ApiResponse response = super.delete( eventId );
-        new MaintenanceActions().removeSoftDeletedEvents();
-
-        return response;
-    }
-
-    public ApiResponse softDelete( String eventId )
-    {
-        ApiResponse response = super.delete( eventId );
-
-        response.validate().statusCode( 200 );
-
-        return response;
-    }
-
-    public void softDelete( List<String> eventIds )
-    {
-        for ( String id : eventIds )
-        {
-            softDelete( id );
-        }
-    }
-
-    public JsonObject createEventBody( String orgUnitId, String programId, String programStageId )
-    {
-        JsonObject event = new JsonObject();
-
-        event.addProperty( "orgUnit", orgUnitId );
-        event.addProperty( "program", programId );
-        event.addProperty( "programStage", programStageId );
-        event.addProperty( "eventDate", "2018-12-01T00:00:00.000" );
-
-        return event;
-    }
 }
