+++ dhis-2/dhis-e2e-test/src/test/java/org/hisp/dhis/tracker/TrackedEntityAttributeValueGenerationTests.java
@@ -53,7 +53,7 @@ import org.junit.jupiter.params.provider.MethodSource;
  * @author Gintare Vilkelyte <vilkelyte.gintare@gmail.com>
  */
 public class TrackedEntityAttributeValueGenerationTests
-    extends TrackerApiTest
+    extends DeprecatedTrackerApiTest
 {
     private RestApiActions trackedEntityAttributeActions;
