+++ dhis-2/dhis-e2e-test/src/test/java/org/hisp/dhis/tracker/TrackerApiTest.java
@@ -27,40 +27,174 @@
  */
 package org.hisp.dhis.tracker;
 
+import java.io.File;
+import java.util.List;
+
 import org.hisp.dhis.ApiTest;
+import org.hisp.dhis.Constants;
 import org.hisp.dhis.actions.LoginActions;
-import org.hisp.dhis.actions.RestApiActions;
+import org.hisp.dhis.actions.MaintenanceActions;
 import org.hisp.dhis.actions.metadata.ProgramActions;
-import org.hisp.dhis.actions.tracker.EventActions;
-import org.hisp.dhis.actions.tracker.TEIActions;
+import org.hisp.dhis.actions.tracker.importer.TrackerActions;
+import org.hisp.dhis.dto.TrackerApiResponse;
+import org.hisp.dhis.helpers.file.FileReaderUtils;
+import org.hisp.dhis.helpers.file.JsonFileReader;
+import org.hisp.dhis.tracker.importer.databuilder.RelationshipDataBuilder;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Tag;
+
+import com.google.gson.JsonObject;
 
 /**
  * @author Gintare Vilkelyte <vilkelyte.gintare@gmail.com>
  */
-@Tag( "category:tracker" )
+@Tag( "tracker" )
+@Tag( "category:tracker_nti" )
 public class TrackerApiTest
     extends ApiTest
 {
-    protected RestApiActions enrollmentActions;
+    protected static final String TRACKER_PROGRAM_STAGE_ID = "nlXNK4b7LVr";
 
-    protected EventActions eventActions;
-
-    protected TEIActions teiActions;
+    protected String TRACKER_PROGRAM_ID = Constants.TRACKER_PROGRAM_ID;
 
     protected LoginActions loginActions;
 
     protected ProgramActions programActions;
 
+    protected TrackerActions trackerActions;
+
     @BeforeAll
     public void beforeTracker()
     {
-        teiActions = new TEIActions();
         loginActions = new LoginActions();
         programActions = new ProgramActions();
-        eventActions = new EventActions();
-        enrollmentActions = new RestApiActions( "/enrollments" );
+        trackerActions = new TrackerActions();
     }
 
+    protected String importEnrollment()
+        throws Exception
+    {
+        JsonObject teiBody = new FileReaderUtils()
+            .readJsonAndGenerateData( new File( "src/test/resources/tracker/importer/teis/teiAndEnrollment.json" ) );
+
+        return trackerActions.postAndGetJobReport( teiBody ).validateSuccessfulImport().extractImportedEnrollments()
+            .get( 0 );
+    }
+
+    protected String importTei()
+        throws Exception
+    {
+        JsonObject teiBody = new FileReaderUtils()
+            .readJsonAndGenerateData( new File( "src/test/resources/tracker/importer/teis/tei.json" ) );
+
+        return trackerActions.postAndGetJobReport( teiBody ).validateSuccessfulImport().extractImportedTeis().get( 0 );
+    }
+
+    protected String importTei( String orgUnit )
+        throws Exception
+    {
+        JsonObject teiBody = new FileReaderUtils()
+            .read( new File( "src/test/resources/tracker/importer/teis/tei.json" ) )
+            .replacePropertyValuesRecursivelyWith( "orgUnit", orgUnit )
+            .get( JsonObject.class );
+
+        return trackerActions.postAndGetJobReport( teiBody ).validateSuccessfulImport().extractImportedTeis().get( 0 );
+    }
+
+    protected List<String> importEvents()
+        throws Exception
+    {
+        JsonObject object = new FileReaderUtils()
+            .read( new File( "src/test/resources/tracker/importer/events/events.json" ) )
+            .replacePropertyValuesWithIds( "event" ).get( JsonObject.class );
+
+        return trackerActions.postAndGetJobReport( object )
+            .validateSuccessfulImport().extractImportedEvents();
+    }
+
+    protected TrackerApiResponse importTeiWithEnrollment( String programId )
+        throws Exception
+    {
+        JsonObject teiWithEnrollment = new FileReaderUtils()
+            .read( new File( "src/test/resources/tracker/importer/teis/teiWithEnrollments.json" ) )
+            .replacePropertyValuesRecursivelyWith( "program", programId )
+            .get( JsonObject.class );
+
+        return trackerActions.postAndGetJobReport( teiWithEnrollment ).validateSuccessfulImport();
+    }
+
+    protected TrackerApiResponse importTeiWithEnrollment( String ouId, String programId )
+        throws Exception
+    {
+        JsonObject teiWithEnrollment = new FileReaderUtils()
+            .read( new File( "src/test/resources/tracker/importer/teis/teiWithEnrollments.json" ) )
+            .replacePropertyValuesRecursivelyWith( "program", programId )
+            .replacePropertyValuesRecursivelyWith( "orgUnit", ouId )
+            .get( JsonObject.class );
+
+        return trackerActions.postAndGetJobReport( teiWithEnrollment ).validateSuccessfulImport();
+    }
+
+    /*
+     * Imports one new TEI with enrollment and event
+     */
+    protected TrackerApiResponse importTeisWithEnrollmentAndEvent( String orgUnit, String programId,
+        String programStageId )
+        throws Exception
+    {
+        JsonObject teiWithEnrollment = new FileReaderUtils()
+            .read( new File( "src/test/resources/tracker/importer/teis/teiWithEnrollmentAndEventsNested.json" ) )
+            .replacePropertyValuesRecursivelyWith( "orgUnit", orgUnit )
+            .replacePropertyValuesRecursivelyWith( "program", programId )
+            .replacePropertyValuesRecursivelyWith( "programStage", programStageId )
+            .get( JsonObject.class );
+
+        TrackerApiResponse response = trackerActions.postAndGetJobReport( teiWithEnrollment );
+
+        response.validateSuccessfulImport();
+        return response;
+    }
+
+    /*
+     * Imports new tracked entities, each having an enrollment and event.
+     */
+    protected TrackerApiResponse importTeisWithEnrollmentAndEvent()
+        throws Exception
+    {
+        JsonObject object = new JsonFileReader(
+            new File( "src/test/resources/tracker/importer/teis/teisWithEnrollmentsAndEvents.json" ) )
+                .replaceStringsWithIds( "Kj6vYde4LHh", "Nav6inZRw1u", "MNWZ6hnuhSw", "PuBvJxDB73z", "olfXZzSGacW",
+                    "ZwwuwNp6gVd" )
+                .get( JsonObject.class );
+
+        return trackerActions.postAndGetJobReport( object )
+            .validateSuccessfulImport();
+    }
+
+    protected TrackerApiResponse importRelationshipBetweenTeis( String teiA, String teiB )
+    {
+        JsonObject payload = new RelationshipDataBuilder()
+            .setFromTrackedEntity( teiA )
+            .setToTrackedEntity( teiB )
+            .setRelationshipType( "xLmPUYJX8Ks" )
+            .array();
+
+        return trackerActions.postAndGetJobReport( payload ).validateSuccessfulImport();
+    }
+
+    protected TrackerApiResponse importRelationshipEnrollmentToTei( String enrollment, String teiB )
+    {
+        JsonObject payload = new RelationshipDataBuilder().setFromEntity( "enrollment", enrollment )
+            .setToTrackedEntity( teiB ).setRelationshipType( "fdc6uOvgoji" ).array();
+
+        return trackerActions.postAndGetJobReport( payload ).validateSuccessfulImport();
+    }
+
+    @AfterEach
+    public void tearDown()
+    {
+        loginActions.loginAsSuperUser();
+        new MaintenanceActions().removeSoftDeletedData();
+    }
 }
