+++ dhis-2/dhis-e2e-test/src/test/java/org/hisp/dhis/tracker/events/EventImportIdSchemeTests.java
@@ -41,7 +41,7 @@ import org.hisp.dhis.dto.ApiResponse;
 import org.hisp.dhis.dto.OrgUnit;
 import org.hisp.dhis.helpers.QueryParamsBuilder;
 import org.hisp.dhis.helpers.file.FileReaderUtils;
-import org.hisp.dhis.tracker.TrackerApiTest;
+import org.hisp.dhis.tracker.DeprecatedTrackerApiTest;
 import org.hisp.dhis.utils.DataGenerator;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -55,7 +55,7 @@ import com.google.gson.JsonObject;
  * @author Gintare Vilkelyte <vilkelyte.gintare@gmail.com>
  */
 public class EventImportIdSchemeTests
-    extends TrackerApiTest
+    extends DeprecatedTrackerApiTest
 {
     private static String OU_NAME = "TA EventsImportIdSchemeTests ou name " + DataGenerator.randomString();
