+++ dhis-2/dhis-e2e-test/src/test/java/org/hisp/dhis/tracker/importer/AtomicModeTests.java
@@ -37,7 +37,7 @@ import org.hisp.dhis.dto.TrackerApiResponse;
 import org.hisp.dhis.helpers.JsonObjectBuilder;
 import org.hisp.dhis.helpers.QueryParamsBuilder;
 import org.hisp.dhis.helpers.file.JsonFileReader;
-import org.hisp.dhis.tracker.TrackerNtiApiTest;
+import org.hisp.dhis.tracker.TrackerApiTest;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
@@ -47,7 +47,7 @@ import com.google.gson.JsonObject;
  * @author Gintare Vilkelyte <vilkelyte.gintare@gmail.com>
  */
 public class AtomicModeTests
-    extends TrackerNtiApiTest
+    extends TrackerApiTest
 {
     @BeforeAll
     public void beforeAll()
