+++ dhis-2/dhis-e2e-test/src/test/java/org/hisp/dhis/tracker/importer/OwnershipTests.java
@@ -36,7 +36,7 @@ import org.hisp.dhis.Constants;
 import org.hisp.dhis.actions.UserActions;
 import org.hisp.dhis.actions.metadata.ProgramActions;
 import org.hisp.dhis.helpers.QueryParamsBuilder;
-import org.hisp.dhis.tracker.TrackerNtiApiTest;
+import org.hisp.dhis.tracker.TrackerApiTest;
 import org.hisp.dhis.tracker.importer.databuilder.EnrollmentDataBuilder;
 import org.hisp.dhis.utils.DataGenerator;
 import org.junit.jupiter.api.BeforeAll;
@@ -51,7 +51,7 @@ import com.google.gson.JsonObject;
  * @author Gintare Vilkelyte <vilkelyte.gintare@gmail.com>
  */
 public class OwnershipTests
-    extends TrackerNtiApiTest
+    extends TrackerApiTest
 {
     String userPassword = Constants.USER_PASSWORD;
