+++ dhis-2/dhis-e2e-test/src/test/java/org/hisp/dhis/tracker/importer/SideEffectsTests.java
@@ -41,7 +41,7 @@ import org.hisp.dhis.dto.ApiResponse;
 import org.hisp.dhis.helpers.JsonObjectBuilder;
 import org.hisp.dhis.helpers.QueryParamsBuilder;
 import org.hisp.dhis.helpers.file.FileReaderUtils;
-import org.hisp.dhis.tracker.TrackerNtiApiTest;
+import org.hisp.dhis.tracker.TrackerApiTest;
 import org.hisp.dhis.tracker.importer.databuilder.TeiDataBuilder;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Disabled;
@@ -56,7 +56,7 @@ import com.google.gson.JsonObject;
  */
 
 public class SideEffectsTests
-    extends TrackerNtiApiTest
+    extends TrackerApiTest
 {
     private String trackerProgramStageId = "PaOOjwLVW23";
