+++ dhis-2/dhis-e2e-test/src/test/java/org/hisp/dhis/tracker/importer/events/EventValidationTests.java
@@ -49,7 +49,7 @@ import org.hisp.dhis.dto.ApiResponse;
 import org.hisp.dhis.dto.TrackerApiResponse;
 import org.hisp.dhis.helpers.QueryParamsBuilder;
 import org.hisp.dhis.helpers.file.FileReaderUtils;
-import org.hisp.dhis.tracker.TrackerNtiApiTest;
+import org.hisp.dhis.tracker.TrackerApiTest;
 import org.hisp.dhis.tracker.importer.databuilder.EventDataBuilder;
 import org.hisp.dhis.tracker.importer.databuilder.TeiDataBuilder;
 import org.junit.jupiter.api.Assumptions;
@@ -66,7 +66,7 @@ import com.google.gson.JsonObject;
  * @author Gintare Vilkelyte <vilkelyte.gintare@gmail.com>
  */
 public class EventValidationTests
-    extends TrackerNtiApiTest
+    extends TrackerApiTest
 {
     private static final String OU_ID = Constants.ORG_UNIT_IDS[0];
