+++ dhis-2/dhis-e2e-test/src/test/java/org/hisp/dhis/tracker/importer/events/EventsDataValueValidationTests.java
@@ -45,7 +45,7 @@ import org.hisp.dhis.dto.ApiResponse;
 import org.hisp.dhis.dto.TrackerApiResponse;
 import org.hisp.dhis.helpers.JsonObjectBuilder;
 import org.hisp.dhis.helpers.QueryParamsBuilder;
-import org.hisp.dhis.tracker.TrackerNtiApiTest;
+import org.hisp.dhis.tracker.TrackerApiTest;
 import org.hisp.dhis.tracker.importer.databuilder.EventDataBuilder;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
@@ -58,7 +58,7 @@ import com.google.gson.JsonObject;
  * @author Gintare Vilkelyte <vilkelyte.gintare@gmail.com>
  */
 public class EventsDataValueValidationTests
-    extends TrackerNtiApiTest
+    extends TrackerApiTest
 {
     private static final String OU_ID = Constants.ORG_UNIT_IDS[0];
